src/identity_echo_api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
PyJWT[crypto]>=2.9.0
httpx>=0.28.0
python-dotenv>=1.0.0
